requirements.txt
tweepy==3.10.0
instaloader==4.7.3
facebook-sdk==3.1.0
python-snapchat==0.1.3
TikTokApi==0.1.0
oauthlib==*
keyring==*
PyQt5~=5.15.11